requirements.txt
google-genai==0.3.0
pdf2image==1.17.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1